fixed more CR change requests by Yam
