+++ app/routers/calendar_grid.py
@@ -181,8 +181,7 @@ def create_weeks(
     num_days: int = len(ndays)
     _weeks = [Week(ndays[i : i + length]) for i in range(0, num_days, length)]
 
-    weeks = add_game_events_to_weeks(_weeks, is_active=False)
-    return weeks
+    return add_game_events_to_weeks(_weeks, is_active=False)
 
 
 def get_month_block(day: Day, n: int = MONTH_BLOCK) -> List[Week]:
