+++ app/static/game_releases.css
@@ -1,0 +1,7 @@
+.game-title {
+  font-size: 1.1em;
+  font-weight: bolder;
+}
+.secondary-title {
+  font-weight: bold;
+}
