+++ app/templates/partials/calendar/feature_settings/game_release_modal.html
@@ -1,4 +1,6 @@
 <!-- upcomingGames Modal -->
+<link href="{{ url_for('static', path='/game_releases.css') }}" rel="stylesheet">
+
 <div
   class="modal modal-lg"
   id="upcoming-games-modal"
@@ -21,6 +23,7 @@
         ></button>
       </div>
       <div class="modal-body">
+        <label for="from-date">From:</label>
         <input
           type="date"
           id="from-date"
@@ -28,6 +31,7 @@
           name="from-date"
           placeholder="From Date"
         />
+        <label for="to-date">To:</label>
         <input
           type="date"
           id="to-date"
